Added new command to paymentDAL, changed inserts into payment tble. Changed Unit to have a metric for YTD that is then added for a query in the UnitDAL

+++ Capstone/Models/Unit.cs
@@ -108,6 +108,6 @@ namespace Capstone.Models
             }
         }
 
-        public int RentCollectedYTD { get; set; }
+        public decimal RentCollectedYTD { get; set; }
     }
 }
